requirements-crewai.txt
# CrewAI Implementation Requirements
# Install with: pip install -r requirements-crewai.txt

# CrewAI framework
crewai==1.7.2
crewai-tools==1.7.2
litellm==1.61.0
# LangChain for LLM integration
langchain==0.1.9
langchain-community==0.0.21

# Core dependencies (shared with custom implementation)
pydantic==2.11.9
pydantic-settings==2.10.1

# YAML support for ADRs
pyyaml==6.0.1

# Utilities
python-dotenv==1.1.1
httpx==0.27.1

# Optional but recommended
colorama==0.4.6  # For colored terminal output
